refactor(channels): 提取 useChannelModelSelection hook 并迁移类型到 types
